add some code to support commonjs
